Move tests to integration test + reorg fixtures
